feat: 웹소켓 시큐리티
1. 웹소켓 시큐리티 적용
2. 로그인과 연결해서 로그인한 사용자만 방생성,방참여 가능

+++ src/main/java/com/service/indianfrog/domain/gameroom/controller/GameRoomController.java
@@ -2,18 +2,26 @@ package com.service.indianfrog.domain.gameroom.controller;
 
 import com.service.indianfrog.domain.gameroom.dto.GameRoomDto;
 import com.service.indianfrog.domain.gameroom.service.GameRoomService;
+import com.service.indianfrog.global.security.UserDetailsImpl;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-@CrossOrigin(origins = "*") // CORS
+import java.util.List;
+
+
 @RestController
 @RequestMapping("/gameRoom")
 public class GameRoomController {
 
-    private GameRoomService gameRoomService;
+    private static final Logger logger = LoggerFactory.getLogger(GameRoomController.class);
+    private final GameRoomService gameRoomService;
     private final SimpMessagingTemplate messagingTemplate;
 
     public GameRoomController(GameRoomService gameRoomService, SimpMessagingTemplate messagingTemplate) {
@@ -21,16 +29,46 @@ public class GameRoomController {
         this.messagingTemplate = messagingTemplate;
     }
 
+
+    @GetMapping("/")
+    public ResponseEntity<List<GameRoomDto>> getAllGameRooms() {
+        List<GameRoomDto> gameRooms = gameRoomService.getAllGameRooms();
+        return ResponseEntity.ok(gameRooms);
+    }
+
+    @GetMapping("/{roomId}")
+    public ResponseEntity<GameRoomDto> getGameRoomById(@PathVariable Long roomId) {
+        GameRoomDto gameRoom = gameRoomService.getGameRoomById(roomId);
+        if (gameRoom != null) {
+            return ResponseEntity.ok(gameRoom);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
+
+
     @PostMapping("/create")
-    public ResponseEntity<GameRoomDto> createGameRoom(@RequestBody GameRoomDto gameRoomDto) {
+    public ResponseEntity<GameRoomDto> createGameRoom(@RequestBody GameRoomDto gameRoomDto, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (userDetails == null) {
+            // 사용자가 로그인하지 않은 경우 접근을 거부합니다.
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        // 게임방 생성 로직을 호출합니다.
         GameRoomDto createdGameRoom = gameRoomService.createGameRoom(gameRoomDto);
-        // GameRoomDto createdGameRoom = gameRoomService.createGameRoom(gameRoomDto, creatorParticipant);  @RequestParam String creatorParticipant
+        // 생성된 게임방 정보를 반환합니다.
         return ResponseEntity.ok(createdGameRoom);
     }
 
     @DeleteMapping("/delete/{roomId}")
-    public ResponseEntity<?> deleteGameRoom(@PathVariable Long roomId) {
+    public ResponseEntity<?> deleteGameRoom(@PathVariable Long roomId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (userDetails == null) {
+            // 사용자가 로그인하지 않은 경우 접근을 거부합니다.
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        // 게임방 삭제 로직을 호출합니다.
         gameRoomService.deleteGameRoom(roomId);
+        // 성공적으로 삭제되었음을 나타내는 응답을 반환합니다.
         return ResponseEntity.ok().build();
     }
 
@@ -52,4 +90,5 @@ public class GameRoomController {
             messagingTemplate.convertAndSend("/topic/gameRoom/" + roomId, gameRoom.getParticipants());
         }
     }
+
 }
